More optimizations
I wrote most of the remaining optimizations that I could think of, including for multiplication/division, exponentiation/root, and log.

I was also able to fix some bugs in parsing of negation and in some of the prior optimizations (specifically optimizations for subtraction).

There is not too much left to do. There are still some more optimizations I want to write (combining of multiplied exponents, combining of summed logs, opt for natlog) and I need to write tests for these and the new optimizations in this commit (such as exponentiation, multiply by zero, etc).