Negation application and tests
Negation application optimization behavior and some new tests. Not all named tests have been finished. The log optimization tests made it clear to me that I want equals(...) to be refined for BinOp. Some operations, like addition and multiplication, are commutative, so if the operations match, albeit in a different order, the equivalence comparison should result true. I imagine that there are other optimizations that could use a refined equality, for example, common subsection computations (if we have expression "(x + 2) * 2 + 1 / (x + 2)", we only need to calculate (x + 2) once and apply it in both locations.

Anyway, the common subsection work can be done in the future. There are doubtless many more optimizations that can be done. However, for this branch specifically I want to finish the tests that I have stubbed out, then merge to main.