Arbitrary log base and removal of explicit maxErr
I realized that the log optimizations should still be valid even if the bases are variable, as long as they are equivalent. Right now I am using equals(...) to test equivalence, but I may need to refine or replace the function with a measure better suited to fit two expressions having the same result (even if through a different structure).

I also was able to clean up the interface for optimization since I don't need to explicitly pass maxErr. The ExpressionParser saves it nicely, which is saved in the ExpressionSolver passed in. Thus, it is already available with a little redirection.

One last change I made was in regards to constants. They need some maxErr to calculate equivalence, and the interface for equals forbids any other parameters. Thus, I passed the maximum error through the constructor, which I am not fully pleased with, but was an acceptable temporary solution.