Multiplication combination optimizations
I wrote two new (very simple) optimizations that I unwittingly required in two of my test cases. They were combining division operands on one or both sides of a multiplication. For example, (1/x)*y = y/x. Or (2/y) * (3/x) = 6/(yx).

I was also able to resolve a small bug in my combineRoots test where I misordered the operands of a multiply (y was Lhs and z was Rhs, but it has been fixed with z as Lhs and y as Rhs). I have already mentioned considerations of some "equivalent" method that would regard the two as equal, but that has yet to come to fruition.