Recursion debugging
There is an interesting problem since I defined the optimizations of division as a convoluted optimization of multiplication. The problem arises since multiplication is smart enough to resolve the convolution which division creates, which brings us into an infinite loop.

For the time being, all the test cases should pass. I could get division's optimizations to end a little cleaner, but now they at least work.

I also added a test in the parser to verify there are not too many closing parentheses.