Optimizations (#3)
Optimization processing for a majority of operations. I was also able to make some bug fixes to the main parsing code, including changing right evaluation to left-evaluation, slightly adjusting the precedence level of operations (specifically unary negation), and handling the error case of too many end parentheses. 

The OptimizationTest file does a pretty good job at tracking all the available optimizations. I can use these going forward to verify there is no regression. 

One snag that I am not fully pleased with is in regards to optimizing Division. It relies on the same optimizations as Multiplication (since they are very similar), but it will easily get caught in an infinite recursion loop if not for the safeguards I installed. However, the safeguards are not very elegant and may prevent some optimizations. I hope to clean this up in the future, though it is not critical for full-functioning behavior.